pyproject.toml
[project]
name = "llmchat"
version = "0.1.0"
description = "Chat based LLM model "
authors = [
    {name = "Pradhap Rajamani",email = "pradhapreee@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "faiss-cpu==1.8.0.post1",
    "langchain==0.2.7",
    "langchain-community==0.2.7",
    "pdfminer.six==20240706",
    "tqdm==4.66.4",
    "waitress==3.0.2"
]
[tool.poetry.scripts]
flask_app = "src.app:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
